refactor: collapse matrix cache thread policies
