+++ tach.toml
@@ -81,8 +81,6 @@ expose = [
     "THREAD_HISTORY_SOURCE_STALE_CACHE",
     "ThreadCacheState",
     "ThreadHistoryResult",
-    "ThreadReadPolicy",
-    "ThreadWritePolicy",
     "normalize_nio_event_for_cache",
     "thread_cache_state_is_usable",
     "thread_history_result",
